Issue/37

close #37

+++ src/main/java/jungle/krafton/AIInterviewMate/controller/RatingController.java
@@ -25,7 +25,7 @@ public class RatingController {
     }
 
     @PostMapping(value = "/{roomIdx}/viewee")
-    public ResponseEntity<PrivateResponseBody> saveRating(@PathVariable int roomIdx, @RequestBody RatingInterviewDto ratingInterviewDto) {
+    public ResponseEntity<PrivateResponseBody> saveRating(@PathVariable Long roomIdx, @RequestBody RatingInterviewDto ratingInterviewDto) {
         ratingService.saveRating(roomIdx, ratingInterviewDto);
         return new ResponseEntity<>(new PrivateResponseBody(StatusCode.OK, null), HttpStatus.OK);
     }
